requirements.txt
flask==3.1.0
python-dotenv==1.0.1
openai==1.68.0
requests==2.32.3
beautifulsoup4==4.13.3
